refactor: convert cleanup endpoint to Vercel cron job
Moves the stale workflow cleanup from a run-domain API endpoint to a
proper Vercel cron handler at /api/cron/cleanup-stale-workflows,
following the existing cleanupStreamChunks pattern:
- noAuth() + CRON_SECRET bearer validation
- System-level (cross-tenant) cleanup
- Runs every 30 minutes via vercel.json cron schedule
- Removes the tenant-scoped run domain route
