+++ packages/agents-core/src/data-access/runtime/workflowExecutions.ts
@@ -68,17 +68,12 @@ export const getWorkflowExecutionByConversation =
 
 export const getStaleWorkflowExecutions =
   (db: AgentsRunDatabaseClient) =>
-  async (params: {
-    scopes: TenantScopeConfig;
-    staleBefore: string;
-    limit?: number;
-  }): Promise<WorkflowExecutionSelect[]> => {
+  async (params: { staleBefore: string; limit?: number }): Promise<WorkflowExecutionSelect[]> => {
     const result = await db
       .select()
       .from(workflowExecutions)
       .where(
         and(
-          eq(workflowExecutions.tenantId, params.scopes.tenantId),
           eq(workflowExecutions.status, 'suspended'),
           lte(workflowExecutions.updatedAt, params.staleBefore)
         )
